server/requirements.txt
Flask
Pillow
requests
opencv-python